update
Former-commit-id: b8fff437e3c090c084116f06ad66dcb24b3fd013

+++ user/llisiw/upgraddsr.c
@@ -64,8 +64,8 @@ void line2cart(int dim       /* number of dimensions */,
     int axis;
  
     for (axis = 0; axis < dim; axis++) {
-	ii[axis] = i%nn[axis];
-	i /= nn[axis];
+      ii[axis] = i%((long) nn[axis]);
+      i /= (long) nn[axis];
     }
 }
